chore: trim ConfigureEvm trait
