+++ crates/evm/src/lib.rs
@@ -17,14 +17,12 @@
 
 extern crate alloc;
 
-use crate::builder::RethEvmBuilder;
 use alloy_consensus::BlockHeader as _;
 use alloy_primitives::{Address, Bytes, B256, U256};
 use reth_primitives_traits::{BlockHeader, SignedTransaction};
 use revm::{Database, Evm, GetInspector};
 use revm_primitives::{BlockEnv, CfgEnvWithHandlerCfg, Env, SpecId, TxEnv};
 
-pub mod builder;
 pub mod either;
 /// EVM environment configuration.
 pub mod env;
@@ -42,26 +40,11 @@ pub mod test_utils;
 
 /// Trait for configuring the EVM for executing full blocks.
 pub trait ConfigureEvm: ConfigureEvmEnv {
-    /// Returns new EVM with the given database
-    ///
-    /// This does not automatically configure the EVM with [`ConfigureEvmEnv`] methods. It is up to
-    /// the caller to call an appropriate method to fill the transaction and block environment
-    /// before executing any transactions using the provided EVM.
-    fn evm<DB: Database>(&self, db: DB) -> Evm<'_, (), DB> {
-        RethEvmBuilder::new(db).build()
-    }
-
     /// Returns a new EVM with the given database configured with the given environment settings,
     /// including the spec id and transaction environment.
     ///
     /// This will preserve any handler modifications
-    fn evm_with_env<DB: Database>(&self, db: DB, evm_env: EvmEnv, tx: TxEnv) -> Evm<'_, (), DB> {
-        let mut evm = self.evm(db);
-        evm.modify_spec_id(evm_env.cfg_env_with_handler_cfg.handler_cfg.spec_id);
-        evm.context.evm.env =
-            Env::boxed(evm_env.cfg_env_with_handler_cfg.cfg_env, evm_env.block_env, tx);
-        evm
-    }
+    fn evm_with_env<DB: Database>(&self, db: DB, evm_env: EvmEnv, tx: TxEnv) -> Evm<'_, (), DB>;
 
     /// Returns a new EVM with the given database configured with `cfg` and `block_env`
     /// configuration derived from the given header. Relies on
@@ -90,27 +73,7 @@ pub trait ConfigureEvm: ConfigureEvmEnv {
     ) -> Evm<'_, I, DB>
     where
         DB: Database,
-        I: GetInspector<DB>,
-    {
-        let mut evm = self.evm_with_inspector(db, inspector);
-        evm.modify_spec_id(evm_env.cfg_env_with_handler_cfg.handler_cfg.spec_id);
-        evm.context.evm.env =
-            Env::boxed(evm_env.cfg_env_with_handler_cfg.cfg_env, evm_env.block_env, tx);
-        evm
-    }
-
-    /// Returns a new EVM with the given inspector.
-    ///
-    /// Caution: This does not automatically configure the EVM with [`ConfigureEvmEnv`] methods. It
-    /// is up to the caller to call an appropriate method to fill the transaction and block
-    /// environment before executing any transactions using the provided EVM.
-    fn evm_with_inspector<DB, I>(&self, db: DB, inspector: I) -> Evm<'_, I, DB>
-    where
-        DB: Database,
-        I: GetInspector<DB>,
-    {
-        RethEvmBuilder::new(db).build_with_inspector(inspector)
-    }
+        I: GetInspector<DB>;
 }
 
 impl<'b, T> ConfigureEvm for &'b T
@@ -118,10 +81,6 @@ where
     T: ConfigureEvm,
     &'b T: ConfigureEvmEnv<Header = T::Header>,
 {
-    fn evm<DB: Database>(&self, db: DB) -> Evm<'_, (), DB> {
-        (*self).evm(db)
-    }
-
     fn evm_for_block<DB: Database>(&self, db: DB, header: &Self::Header) -> Evm<'_, (), DB> {
         (*self).evm_for_block(db, header)
     }
@@ -142,14 +101,6 @@ where
         I: GetInspector<DB>,
     {
         (*self).evm_with_env_and_inspector(db, evm_env, tx_env, inspector)
-    }
-
-    fn evm_with_inspector<DB, I>(&self, db: DB, inspector: I) -> Evm<'_, I, DB>
-    where
-        DB: Database,
-        I: GetInspector<DB>,
-    {
-        (*self).evm_with_inspector(db, inspector)
     }
 }
